+++ js/image-modify.js
@@ -1,0 +1,36 @@
+const buttonScaleControlValue = document.querySelector('.scale__control--value');
+const buttonScaleControlSmaller = document.querySelector('.scale__control--smaller');
+const buttonScaleControlBigger = document.querySelector('.scale__control--bigger');
+
+const imageUploadPreview = document.querySelector('.img-upload__preview');
+
+function onClickButtonScaleControlSmaller () {
+  if (buttonScaleControlValue.value > 25 && buttonScaleControlValue.value < 100) {
+    buttonScaleControlValue.value -= 25;
+  }
+}
+
+function onClickButtonScaleControlBigger () {
+  if (buttonScaleControlValue.value < 100 && buttonScaleControlValue.value > 25) {
+    buttonScaleControlValue.value += 25;
+  }
+}
+
+function onChangeScaleControlValue () {
+  const scale = buttonScaleControlValue.value / 100;
+  imageUploadPreview.attributeStyleMap.set('transform', `scale(${scale})`);
+}
+
+function addEventOnElementsWrapper () {
+  buttonScaleControlSmaller.addEventListener('click',onClickButtonScaleControlSmaller);
+  buttonScaleControlBigger.addEventListener('click',onClickButtonScaleControlBigger);
+  buttonScaleControlValue.addEventListener('change',onChangeScaleControlValue);
+}
+
+function removeEventOnElementsWrapper () {
+  buttonScaleControlSmaller.removeEventListener('click',onClickButtonScaleControlSmaller);
+  buttonScaleControlBigger.removeEventListener('click',onClickButtonScaleControlBigger);
+  buttonScaleControlValue.removeEventListener('change',onChangeScaleControlValue);
+}
+
+export {addEventOnElementsWrapper, removeEventOnElementsWrapper};
